pyproject.toml
[tool.poetry]
name = "mtgdb"
version = "0.0.1"
description = "Magic Card Database and Tools"
authors = ["Pasha Sadikov <sigmat@proton.me>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.15"
magicinvoke = "^2.4.6"
pydantic = "^1.10.4"
tabulate = "^0.9.0"
urwid = "^2.1.2"

[tool.poetry.group.projects.dependencies]
mtg-db = {path = "projects/mtg-db", develop = true}
cursed-inventory = {path = "projects/cursed-inventory", develop = true}

[tool.poetry.group.dev.dependencies]
ipython = "^8.8.0"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "pep440"
bump = true
dirty = true

# --------------------------------------------------
[build-system]
requires = [
    "poetry-core>=1.0.0",
    "poetry-dynamic-versioning"
]
build-backend = "poetry.core.masonry.api"
